feat: add configuration and schema management for Mise and Aqua registries
- Introduced a new `config.py` file to centralize OS and architecture mapping, along with registry base URLs.
- Created `schemas.py` to define data classes for `AquaAsset` and `MiseToolInfo`.
- Refactored `registry.py` to utilize the new configuration and schema definitions, improving code organization and maintainability.
- Updated import paths in `main.py` to reflect the new structure.
